Вынес в отдельный файл AppBasedSdkUiProvider
