+++ TinkoffASDKCore/TinkoffASDKCore/Domain/3DS/ThreeDSDeviceInfoProvider/ThreeDSDeviceInfoProvider.swift
@@ -30,32 +30,6 @@ public extension IThreeDSDeviceInfoProvider {
     }
 }
 
-/// Возвращает информацию для 3DS App Based SDK транзакции
-public protocol IAppBasedSdkUiProvider {
-    /// Тип интерфейса Native/HTML/Both через который пойдет транзакция
-    func sdkInterface() -> TdsSdkInterface
-    /// Тип ui-ая для проверки
-    func sdkUiTypes() -> [TdsSdkUiType]
-}
-
-public struct AppBasedSdkUiProvider: IAppBasedSdkUiProvider {
-    let prefferedInterface: TdsSdkInterface
-    let prefferedUiTypes: [TdsSdkUiType]
-
-    public init(prefferedInterface: TdsSdkInterface, prefferedUiTypes: [TdsSdkUiType]) {
-        self.prefferedInterface = prefferedInterface
-        self.prefferedUiTypes = prefferedUiTypes
-    }
-
-    public func sdkInterface() -> TdsSdkInterface {
-        prefferedInterface
-    }
-
-    public func sdkUiTypes() -> [TdsSdkUiType] {
-        prefferedUiTypes
-    }
-}
-
 final class ThreeDSDeviceInfoProvider: IThreeDSDeviceInfoProvider {
     private let languageProvider: ILanguageProvider
     private let urlBuilder: IThreeDSURLBuilder
